MAS-107 Rename tests
